+++ Sources/SwifterSwift/UIKit/UIViewExtensions.swift
@@ -232,7 +232,6 @@ public extension UIView {
             frame.origin.y = newValue
         }
     }
-
 }
 
 // MARK: - Methods
@@ -593,7 +592,54 @@ public extension UIView {
     func ancestorView<T: UIView>(withClass name: T.Type) -> T? {
         return ancestorView(where: { $0 is T }) as? T
     }
+}
 
+// MARK: - Constraints
+public extension UIView {
+    /// SwifterSwift: Search constraints until we find one for the given view
+    /// and attribute. This will enumerate ancestors since constraints are
+    /// always added to the common ancestor.
+    ///
+    /// - Parameter attribute: the attribute to find
+    /// - Parameter at: the view to find
+    /// - Returns: matching constraint
+    func findConstraint(attribute: NSLayoutConstraint.Attribute, for view: UIView) -> NSLayoutConstraint? {
+        let constraint = constraints.first {
+            ($0.firstAttribute == attribute && $0.firstItem as? UIView == view) ||
+            ($0.secondAttribute == attribute && $0.secondItem as? UIView == view)
+        }
+        return constraint ?? superview?.findConstraint(attribute: attribute, for: view)
+    }
+
+    /// SwifterSwift: First width constraint for this view
+    var widthConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .width, for: self)
+    }
+
+    /// SwifterSwift: First height constraint for this view
+    var heightConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .height, for: self)
+    }
+
+    /// SwifterSwift: First leading constraint for this view
+    var leadingConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .leading, for: self)
+    }
+
+    /// SwifterSwift: First trailing constraint for this view
+    var trailingConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .trailing, for: self)
+    }
+
+    /// SwifterSwift: First top constraint for this view
+    var topConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .top, for: self)
+    }
+
+    /// SwifterSwift: First bottom constraint for this view
+    var bottomConstraint: NSLayoutConstraint? {
+        findConstraint(attribute: .bottom, for: self)
+    }
 }
 
 #endif